requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
discord.py==1.3.3
docker==4.2.0
Flask==1.1.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.5
pylint==2.4.4
python-dotenv==0.13.0
requests==2.23.0
selenium==3.141.0
six==1.14.0
splinter==0.13.0
timeout-decorator==0.4.1
typed-ast==1.4.1
urllib3==1.25.9
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.11.2
yarl==1.4.2
